+++ app/src/main/java/com/app/arcabyolimpo/presentation/navigation/NavGraph.kt
@@ -29,7 +29,8 @@ import com.app.arcabyolimpo.presentation.screens.passwordregisteration.PasswordR
 import com.app.arcabyolimpo.presentation.screens.passwordregisteration.PasswordRegistrationSuccessScreen
 import com.app.arcabyolimpo.presentation.screens.product.addProduct.ProductAddScreen
 import com.app.arcabyolimpo.presentation.screens.product.updateProduct.ProductUpdateScreen
-import com.app.arcabyolimpo.presentation.screens.product.list.ProductsListRoute
+import com.app.arcabyolimpo.presentation.screens.product.list.ProductsListScreen
+import com.app.arcabyolimpo.presentation.screens.product.list.ProductsUiState
 import com.app.arcabyolimpo.presentation.screens.product.productDetail.ProductDeleteTestScreen
 import com.app.arcabyolimpo.presentation.screens.productbatches.productBatchDetail.ProductBatchDetailScreen
 import com.app.arcabyolimpo.presentation.screens.productbatches.productBatchModify.ProductBatchModifyScreen
@@ -135,6 +136,10 @@ sealed class Screen(
 
     object ProductAdd : Screen("product/add")
 
+    object ProductDetail : Screen("product_detail/{productId}") {
+        fun createRoute(productId: String) = "product_detail/$productId"
+    }
+
     object ProductDeleteTest : Screen("test_delete_product")
 
     object SupplyUpdate : Screen("supply/update/{idSupply}") {
@@ -144,6 +149,8 @@ sealed class Screen(
     object ProductUpdate : Screen("product/update/{idProduct}") {
         fun createRoute(idProduct: String) = "product/update/$idProduct"
     }
+
+    object ProductList : Screen("product/")
 
 /**
  * Composable function that defines the main navigation graph of the app.
@@ -167,7 +174,7 @@ fun ArcaNavGraph(
      */
     LaunchedEffect(sessionManager) {
         sessionManager.sessionExpired.collect {
-            navController.navigate(Screen.Login.route) {
+            navController.navigate(Screen.ProductList.route) {
                 popUpTo(0) { inclusive = true }
                 launchSingleTop = true
             }
@@ -180,7 +187,7 @@ fun ArcaNavGraph(
     NavHost(
         navController = navController,
         // TODO: Cambiar a Screen.Splash.route cuando acabe
-        startDestination = Screen.Splash.route,
+        startDestination = Screen.ProductList.route,
         modifier = modifier,
     ) {
         /** Splash Screen */
@@ -661,10 +668,14 @@ fun ArcaNavGraph(
             )
         }
 
-        composable("products_list") {
-            ProductsListRoute(
-                onProductClick = { /* ir a detalle si quieres */ },
-                onBackClick = { navController.popBackStack() },
+        composable(Screen.ProductList.route) {
+            ProductsListScreen(
+                state = ProductsUiState(),
+                onSearchChange = { },
+                onBackClick= { navController.popBackStack() },
+                onDetailClick= { productId-> navController.navigate(Screen.ProductDetail.createRoute(productId)) },
+                onAddClick= { navController.navigate(Screen.ProductAdd.route) },
+                onFilterClick= {  },
             )
         }