+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [0.1.0] - 18/11/2025
 
 ### Added
+- ProductList screen to navgraph
 - Modify product implementation
 
 ### Fixed